# openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "MCP – Microsoft Learn Catalog",
    "version": "1.0.1"
  },
  "servers": [
    {
      "url": "https://mcp-learn-catalog-app-azevddenc4cmdzdp.westeurope-01.azurewebsites.net"
    }
  ],
  "paths": {
    "/searchCatalog": {
      "get": {
        "operationId": "searchCatalog",
        "summary": "Search Microsoft Learn catalog",
        "parameters": [
          {
            "name": "locale",
            "in": "query",
            "description": "Idioma de los resultados (ej. es-es, en-us)",
            "schema": { "type": "string" }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Palabra clave para buscar en el catálogo",
            "schema": { "type": "string" }
          },
          {
            "name": "type",
            "in": "query",
            "description": "Lista de tipos separados por coma (ej. modules,learningPaths)",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Datos del catálogo",
            "content": {
              "application/json": {
                "schema": { "type": "object" }
              }
            }
          }
        }
      }
    },
    "/streamCatalog": {
      "get": {
        "operationId": "streamCatalog",
        "summary": "Stream Microsoft Learn catalog",
        "x-ms-sse": true,
        "parameters": [
          { "name": "locale", "in": "query", "schema": { "type": "string" } },
          { "name": "search", "in": "query", "schema": { "type": "string" } },
          { "name": "type", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Streamed catalog data",
            "content": {
              "application/json": {
                "schema": { "type": "object" }
              }
            }
          }
        }
      }
    },
    "/sseCatalog": {
      "get": {
        "operationId": "sseCatalog",
        "summary": "SSE Microsoft Learn catalog",
        "x-ms-sse": true,
        "parameters": [
          { "name": "locale", "in": "query", "schema": { "type": "string" } },
          { "name": "search", "in": "query", "schema": { "type": "string" } },
          { "name": "type", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "SSE streamed catalog data",
            "content": {
              "text/event-stream": {
                "schema": { "type": "string" }
              }
            }
          }
        }
      }
    }
  }
}